# The Big Push
## API updates
I've modified the layouts a bit and added some comments to the api
calls. I've also added a call for the current weather which is part
of the problem. I think I have a solution by creating a second function
to append data to the projected container.

## Appending DOOOOOOM
So currently this appends everything line 82-106.
The loop doesn't seem to like the doing the things so that needs to
get resolved. I feel like its in how it's trying to read the data though
so I think adding an extra function is the easiest way to go.

## So, What next?
Well, next up is the places widget. That should be done right quick.